refactor: crews 관련 타입들(request, response 제외)을 types/crew로 분리
